run uploads in seperate thread
sync will now gather a file list to feed to the upload thread so the UI
remains operable during transfer.

i currently have no way to interrupt a transfer to stop it, so for now
once it starts, it finishes (if it can)

also let's push the timeout to 3 seconds, why not be generous?